+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 All notable changes to MoGAAAP will be documented in this file.
 
+## [1.2.0 - 2026-03-06]
+
+### Added
+- Add validation check for YAHS configuration (#122).
+
+### Changed
+- No longer re-run Helixer for all samples if only one has to be re-run (#123).
+- Update the pipeline in such a way that it is truly modular for the QA steps (#125).
+
 ## [1.1.0 - 2025-12-05]
 
 ### Added